requirements.txt
openai>=1.3.7
jinja2
markdown
reportlab
python-dotenv
pandas  
